project/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic[email]==2.6.1
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
